criação da parte das informções do banco de dados
