+++ .gitignore
@@ -5,6 +5,7 @@
 node_modules
 *.sketch
 *.fig
+*.vscode
 
 # Avoid accidental XD upload if you convert the design file
 ###############################################
