+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Menumobile"]
+  "cSpell.words": ["apoios", "arrecadado", "Menumobile"]
 }
