+++ index.html
@@ -69,11 +69,11 @@
       <section class="support-information">
         <div class="cards-information">
           <div data-information="values" class="values-information">
-            <h2 data-information-title="values">R$ 89.929</h2>
-            <p>de R$ 200.000,00 apoio</p>
+            <h2 data-information-title="values">R$ 0</h2>
+            <p class="information-text">de R$ 200.000,00</p>
           </div>
           <div data-information="backed" class="backed-information">
-            <h2 data-information-title="backed">5,007</h2>
+            <h2 data-information-title="backed">0</h2>
             <p>Total de apoios</p>
           </div>
           <div data-information="days" class="days-information">
