+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
-    "axios": "^1.4.0"
+    "axios": "^1.4.0",
+    "date-fns": "^2.30.0"
   }
 }
